Adding unmount event for nova bridge
When Nova bridge is unmounted, even though the dom is removed, the virtual tree created by Nova using `ReactDOM.render()` does not get removed.
`render` method create a new tree will not be reconciled properly and need to be cleaned up manually.

This event will help the users to handle the unmount and do the cleanup manually when Nova unmounts.